Consolidate apis/<group>/v1 packages into a single apis/v1 package
The per-group k8s-style layout (apis/subscriptions/v1, apis/timeline/v1,
apis/feedentries/v1, apis/viewer/v1) added import-aliasing overhead for
little benefit at this API's current size (5 endpoints, no independent
versioning needs), and PaginationMeta didn't cleanly belong to any single
group. Collapse into one apis/v1 package, split by file per resource
instead of by Go package; internal/api/ now imports it once as apiv1.
