+++ apis/v1/feedentries.go
@@ -1,5 +1,3 @@
-// Package v1 contains the wire types for the feedentries API group, at
-// version v1. See apis/subscriptions/v1 for the rationale on this layout.
 package v1
 
 import "time"
