+++ apis/v1/v1.go
@@ -1,0 +1,4 @@
+// Package v1 contains the wire types for the Seymour API: plain structs
+// with JSON tags, no internal/ dependencies, safe for external callers to
+// import.
+package v1
