+++ internal/api/api.go
@@ -16,7 +16,7 @@ import (
 	"go.temporal.io/sdk/client"
 	"golang.org/x/oauth2"
 
-	feedentriesv1 "github.com/jdholdren/seymour/apis/feedentries/v1"
+	apiv1 "github.com/jdholdren/seymour/apis/v1"
 	"github.com/jdholdren/seymour/internal/seymour"
 )
 
@@ -68,7 +68,7 @@ type Server struct {
 	*http.Server
 
 	fetchClient    *http.Client
-	entryRespCache *lru.Cache[string, feedentriesv1.FeedEntryResp]
+	entryRespCache *lru.Cache[string, apiv1.FeedEntryResp]
 
 	feeds    seymour.FeedService
 	timeline seymour.TimelineService
@@ -93,7 +93,7 @@ func NewServer(
 ) *Server {
 	var (
 		r        = errRouter{Router: mux.NewRouter()}
-		cache, _ = lru.New[string, feedentriesv1.FeedEntryResp](1024)
+		cache, _ = lru.New[string, apiv1.FeedEntryResp](1024)
 	)
 
 	srvr := Server{
